move Sql related functions out of functions.py
SQL related tools now reside within utils/sql. These are split between sql.py and db.py, depending on their purpose.
