+++ cogs/owner.py
@@ -14,6 +14,7 @@ from pathlib import Path
 import discord
 import functions
 from utils.checks import checks
+from utils.sql.db import backups
 from discord.ext import commands, tasks
 
 import cogs.fancyEmbeds as fEmbeds
@@ -225,7 +226,7 @@ class Owner(commands.Cog):
         if os.path.isfile("resources/backups/tempbackupfile.db"):
             await ctx.send("A backup is already in the process of being made! Please wait a moment before trying this again")
             return()
-        await functions.make_backup(self.connection, self.bot.kept_backups)
+        await backups.make_backup(self.connection, self.bot.kept_backups)
         root_directory = Path('resources/backups')
         #functions in f-string gets size, count of everything in "backups" folder, 1 is subtracted from count because of gitkeep
         await ctx.send(f"Sounds good! I made a backup of your database. Currently, your {(len(os.listdir('resources/backups')))-1} backup(s) take up {round((sum(f.stat().st_size for f in root_directory.glob('**/*') if f.is_file())/1000),2)} kilobytes of space")
@@ -322,7 +323,7 @@ class Owner(commands.Cog):
         if os.path.isfile("resources/backups/tempbackupfile.db"):
             logging.warning("Unable to automatically create backup, database backup is already in process. If this problem persists, please contact SushiInYourFace")
             return() #should probably log this occurance, as it may signal something going wrong
-        await functions.make_backup(self.connection, self.bot.kept_backups)
+        await backups.make_backup(self.connection, self.bot.kept_backups)
         logging.info("Database backup created.")
 
     @auto_backup.before_loop
